libs/botte-dynamodb-tasks/pyproject.toml
[project]
name = "botte-dynamodb-tasks"
description = "Botte Monorepo: Botte DynamoDB Tasks"
authors = [
    {name = "puntonim", email = "puntonim@gmail.com"}
]
readme = "README.md"
license = "no license"
version = "1.0.0"
requires-python = ">=3.10,<4.0"  # <4 required by AWS utils (required by Lambda powertools, which is an extra optional).
dependencies = [
    "boto3 (>=1.27.1,<2)",
    "svix-ksuid (>=0.6.2,<0.7.0)",
    "datetime-utils @ git+https://github.com/puntonim/utils-monorepo#subdirectory=datetime-utils",
    "aws-utils @ git+https://github.com/puntonim/utils-monorepo#subdirectory=aws-utils"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipdb (>=0.13.13)",
    # Python linter and formatter.
    "ruff (==0.14.2)",  # Must be the same as in .pre-commit-config.yaml.
]
test = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-xdist[psutil] (>=3.8.0,<4.0.0)",
]

[tool.ruff]
line-length = 88  # Default.
extend-exclude = ["docs"]

[tool.ruff.format]
quote-style = "double"  # Default.

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = [
    # E501 Line too long: let the ruff formatter take care long lines.
    "E501",
]
